refactor: adjust existing method names for new features
