+++ scripts/users_scrapper.ts
@@ -1,0 +1,3 @@
+export async function runScrapper() {
+  return 'Test'
+}
