+++ start/scheduler.ts
@@ -1,5 +1,5 @@
 import scheduler from 'adonisjs-scheduler/services/main'
-import { runScrapper } from '../scripts/scrapper.js'
+import { runScrapper } from '../scripts/menu_scrapper.js'
 import logger from '@adonisjs/core/services/logger'
 
 scheduler
